requirements.txt
playwright>=1.41.0
pydantic>=2.5.0
google-genai>=0.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
matplotlib>=3.8.0
numpy>=1.26.0
python-dotenv>=1.0.0
openai>=1.10.0
openpyxl>=3.1.2
langgraph
langchain-core
